+++ packages/metrics/src/Metric/index.ts
@@ -1,5 +1,4 @@
 export { Metric } from "./Metric";
 export type { MetricDataReference } from "./MetricDataReference";
 export type { MetricDefinition } from "./MetricDefinition";
-export type { MetricCategory } from "./MetricCategory";
-export type { MetricLevel, MetricLevelSet } from "./MetricLevel";
+export type { Category, CategorySet } from "./Category";