+++ packages/ui-components/src/stories/mockMetricCatalog.ts
@@ -23,8 +23,8 @@ const testMetricDefs: MetricDefinition[] = [
     dataReference: {
       providerId: "apple_stock",
     },
-    thresholds: [100, 300, 400, 500],
-    levelSetId: "5_risk_categories",
+    categoryThresholds: [100, 300, 400, 500],
+    categorySetId: "5_risk_categories",
   },
   {
     id: MetricId.PI,
@@ -43,8 +43,8 @@ const testMetricDefs: MetricDefinition[] = [
       providerId: "mock",
       startDate: "2020-01-01",
     },
-    thresholds: [40, 100],
-    levelSetId: "cases_mock",
+    categoryThresholds: [40, 100],
+    categorySetId: "cases_mock",
   },
   {
     id: MetricId.MOCK_CASES_NO_EXTENDED_NAME,
@@ -54,8 +54,8 @@ const testMetricDefs: MetricDefinition[] = [
       providerId: "mock",
       startDate: "2020-01-01",
     },
-    thresholds: [10, 100],
-    levelSetId: "cases_mock",
+    categoryThresholds: [10, 100],
+    categorySetId: "cases_mock",
   },
   {
     id: MetricId.PASS_FAIL,
@@ -65,10 +65,8 @@ const testMetricDefs: MetricDefinition[] = [
       providerId: "static",
       value: 0,
     },
-    categories: [
-      { color: "#d32f2f", value: 0, label: "Fail" },
-      { color: "#4caf50", value: 1, label: "Pass" },
-    ],
+    categorySetId: "pass_fail",
+    categoryValues: [0, 1],
   },
 ];
 
@@ -78,10 +76,10 @@ export const dataProviders = [
   new AppleStockDataProvider(),
 ];
 
-const metricLevelSets = [
+const metricCategorySets = [
   {
     id: "cases_mock",
-    levels: [
+    categories: [
       { id: "low", name: "low", color: theme.palette.severity[100] },
       {
         id: "medium",
@@ -96,11 +94,19 @@ const metricLevelSets = [
         description: "indicates high case incidence.",
       },
     ],
-    defaultLevel: { id: "unknown", color: theme.palette.border.default },
+    defaultCategory: { id: "unknown", color: theme.palette.border.default },
+  },
+  {
+    id: "pass_fail",
+    categories: [
+      { id: "fail", color: "#d32f2f", name: "Fail" },
+      { id: "pass", color: "#4caf50", name: "Pass" },
+    ],
+    defaultCategory: { id: "unknown", color: theme.palette.border.default },
   },
   {
     id: "5_risk_categories",
-    levels: [
+    categories: [
       { id: "low", name: "Low", color: theme.palette.severity[100] },
       {
         id: "medium",
@@ -127,12 +133,12 @@ const metricLevelSets = [
         description: "Critical",
       },
     ],
-    defaultLevel: { id: "unknown", color: theme.palette.border.default },
+    defaultCategory: { id: "unknown", color: theme.palette.border.default },
   },
 ];
 
 export const metricCatalog = new MetricCatalog(testMetricDefs, dataProviders, {
-  metricLevelSets,
+  categorySets: metricCategorySets,
 });
 
 // Exporting a second metric catalog to confirm that the closest